# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurPoint API",
    "version": "0.3.0",
    "description": "Tour marketplace API - operators, tours, group formation pricing, bookings, reviews, deals, and the smart trip planner. Most write operations and anything scoped to \"the logged-in account\" require a Bearer JWT; identity is always derived from the verified token, never from the request body. Sprint 3: all responses carry standard security headers (helmet); /api/auth/login and /api/auth/signup are rate-limited (20 requests/15min per IP) and every other route shares a broader 300 requests/15min ceiling; in production CORS is restricted to the configured frontend origin via CORS_ORIGIN."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/tours": {
      "post": {
        "summary": "Create a tour - requires login as an operator; operator_id is derived from the token and is not accepted in the request body. `features` is an optional comma-separated list of slugs from: breakfast, evening_tea, guide, road_games, hotel_stay - drives the homepage's feature-tag filter chips. `vehicle_features` is an optional comma-separated list from: ac, wifi, luggage, charging - drives the homepage's vehicle-filter chips (falls back to the operator's own vehicle_features, if any, when a tour doesn't set its own).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "title": "Quba Tebiet Turu",
                "price": 50,
                "date": "2026-09-10",
                "features": "breakfast,guide",
                "vehicle_features": "ac,wifi"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Missing title/price/date, non-positive price, or min_participants greater than max_participants"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "403": {
            "description": "This account has no operator profile yet"
          }
        }
      },
      "get": {
        "summary": "Search/list tours",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tours"
          }
        }
      }
    },
    "/tours/compare": {
      "get": {
        "summary": "Compare 2-3 tours side by side",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1,2,3"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tours"
          },
          "400": {
            "description": "Must be 2-3 ids"
          }
        }
      }
    },
    "/tours/popular": {
      "get": {
        "summary": "\"Populyar turlar\" for the homepage - tours ranked by (times favorited + times viewed), most popular first. Must be requested before /tours/{id} in routing terms, but that's an implementation detail.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Max results, 1-12 (default 6)"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tours, most popular first"
          }
        }
      }
    },
    "/tours/{id}": {
      "get": {
        "summary": "Get one tour (includes discounted_price if an active deal exists). Counts as a page view at most once per visitor per 30 minutes; the owning operator's own views are never counted (an optional Bearer token is used only to recognise the owner).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tour"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update a tour - must be the owning operator. Partial update: only fields present in the body are changed",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "price": 60,
                "date": "2026-09-15"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tour"
          },
          "400": {
            "description": "Invalid field value (e.g. empty title, non-positive price, min_participants greater than max_participants)"
          },
          "403": {
            "description": "Not the owning operator"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a tour - must be the owning operator. Blocked if bookings already exist unless ?force=true; cascades cleanup to bookings, group_formations, reviews, and last_minute_deals for that tour",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Delete anyway even if the tour has existing bookings"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "description": "Not the owning operator"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Tour has existing bookings - pass ?force=true to delete anyway"
          }
        }
      }
    },
    "/tours/{id}/photo": {
      "post": {
        "summary": "Upload/replace a tour's photo - multipart/form-data with a single 'photo' image file field (max 5MB). Must be the owning operator; the tour must already exist (create it first via POST /tours).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tour, with the new photo_url"
          },
          "400": {
            "description": "No file uploaded, or the file isn't an image"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "403": {
            "description": "Not the owning operator"
          },
          "404": {
            "description": "Tour not found"
          }
        }
      }
    },
...
  }
}